core/rag/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
chromadb==0.5.0
httpx==0.27.0
pydantic-settings==2.4.0
python-multipart==0.0.9
python-docx==1.1.0
pymupdf==1.24.0
openpyxl==3.1.5
